+++ lib/flutter_statusbarcolor.dart
@@ -11,4 +11,12 @@ class FlutterStatusbarcolor {
       _channel.invokeMethod('setstatusbarcolor', {'color': color.value});
   static Future setNavigationBarColor(Color color) =>
       _channel.invokeMethod('setnavigationbarcolor', {'color': color.value});
+  static Future setStatusBarWhiteForeground(bool useWhiteForeground) =>
+      _channel.invokeMethod('setstatusbarwhiteforeground', {'whiteForeground': useWhiteForeground});
+  static Future setNavigationBarWhiteForeground(bool useWhiteForeground) =>
+      _channel.invokeMethod('setnavigationbarwhiteforeground', {'whiteForeground': useWhiteForeground});
+}
+
+bool useWhiteForeground(Color color) {
+  return 1.05 / (color.computeLuminance() + 0.05) > 4.5;
 }